+++ src/library/model_runner-backend-cpu.adb
@@ -1,5 +1,6 @@
 with Ada.Unchecked_Deallocation;
 
+with Model_Runner.Kernels;
 with Model_Runner.Platform;
 with Model_Runner.Quantization;
 
@@ -831,4 +832,6 @@ begin
      (Model_Runner.Platform.Wide_Vectors);
    Model_Runner.Quantization.Integers.Use_Deep_Rows
      (Model_Runner.Platform.Byte_Products);
+   Model_Runner.Kernels.Use_Wide_Dots
+     (Model_Runner.Platform.Wide_Vectors);
 end Model_Runner.Backend.CPU;